refactor: refactor the duplicate code about charts (#838)
